+++ reservacionVuelo/Screens/SignUp.js
@@ -1,13 +1,8 @@
 import React, {Component} from 'react';
-import {
-  StyleSheet,
-  Text,
-  TextInput,
-  TouchableHighlight,
-  View,
-} from 'react-native';
+import {Text, TextInput, TouchableHighlight, View} from 'react-native';
 import CheckBox from '@react-native-community/checkbox';
 import Icon from 'react-native-vector-icons/EvilIcons';
+import styles from '../components/styles';
 
 export class SignUp extends Component {
   constructor() {
@@ -65,58 +60,3 @@ export class SignUp extends Component {
     );
   }
 }
-
-const styles = StyleSheet.create({
-  mainContainer: {
-    paddingHorizontal: 30,
-    flex: 1,
-  },
-  titleStyle: {
-    fontSize: 40,
-    fontWeight: '700',
-    color: '#5974F5',
-  },
-  textStyle: {
-    color: '#868686',
-  },
-  bottomTextStyle: {
-    textAlign: 'center',
-    color: '#A3ACB9',
-    fontSize: 15,
-    paddingVertical: 10,
-  },
-  textInputTitle: {
-    fontSize: 15,
-    paddingBottom: 5,
-  },
-  textInputStyle: {
-    borderWidth: 1,
-    borderBottomColor: 'black',
-    width: 350,
-    height: 50,
-  },
-  textInputSpacing: {
-    paddingVertical: 10,
-  },
-  checkBoxAlignment: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    right: 5,
-  },
-  buttonStyle: {
-    backgroundColor: '#5974F5',
-    borderRadius: 5,
-    height: 50,
-    alignItems: 'center',
-    paddingTop: 15,
-  },
-  buttonTextStyle: {
-    color: 'white',
-    textAlign: 'center',
-    fontSize: 15,
-  },
-  hyperlinkText: {
-    color: '#5974F5',
-    textDecorationLine: 'underline',
-  },
-});
